tf_flowers-classification-main/flask-image-api/requirements.txt
Flask==3.1.0
Pillow==11.1.0
gunicorn==23.0.0
flask-cors==4.0.0
flask-limiter==3.5.0